[OPIK-264] Minor fixes (#461)
* [OPIK-264] Minor fixes

* Added mapper class

* Removing remaining map logic

* Address PR review feedback
